requirements.txt
altgraph==0.17
certifi==2020.4.5.2
chardet==3.0.4
chart-studio==1.1.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
idna==2.9
montydb==2.1.1
numpy==1.18.5
pandas==1.0.4
pefile==2019.4.18
plotly==4.8.1
PyInstaller==3.6
pymongo==3.10.1
pymongo-inmemory==0.1.3
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.23.0
retrying==1.3.3
six==1.15.0
smmap==3.0.4
urllib3==1.25.9
